+++ graphdatascience/model/node_classification_model.py
@@ -1,13 +1,13 @@
-from typing import Any
+from typing import Any, Dict, List
 
 from pandas import Series
 
 from ..graph.graph_object import Graph
 from ..graph.graph_type_check import graph_type_check
-from .model import Model
+from .pipeline_model import PipelineModel
 
 
-class NCModel(Model):
+class NCModel(PipelineModel):
     def _query_prefix(self) -> str:
         return "CALL gds.beta.pipeline.nodeClassification.predict."
 
@@ -22,3 +22,10 @@ class NCModel(Model):
     @graph_type_check
     def predict_write_estimate(self, G: Graph, **config: Any) -> "Series[Any]":
         return self._estimate_predict("write", G.name(), config)
+
+    def classes(self) -> List[int]:
+        return self._list_info()["modelInfo"][0]["classes"]  # type: ignore
+
+    def feature_properties(self) -> List[str]:
+        features: List[Dict[str, Any]] = self._list_info()["modelInfo"][0]["pipeline"]["featureProperties"]
+        return [f["feature"] for f in features]